requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-tastypie==0.14.5
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
openai==0.27.8
psycopg2==2.9.6
python-dateutil==2.8.2
python-mimeparse==1.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
